Rename expectation.py to result.py
