+++ src/qibochem/measurement/result.py
