+++ tests/test_basis_rotation.py
@@ -7,10 +7,8 @@ import pytest
 from qibo import Circuit, gates, models
 
 from qibochem.ansatz import basis_rotation
-from qibochem.driver.molecule import Molecule
-from qibochem.measurement.expectation import expectation
-
-# from qibo.optimizers import optimize
+from qibochem.driver import Molecule
+from qibochem.measurement import expectation
 
 
 def test_unitary():
